Initial commit
So far I haven't gotten any of the puzzles, mostly because I'm
running out of time.

Messing around with all the test values has been a pain and I've
lost code that may be useful already so I'm making a small
repository for the stuff I'm doing while working on these projects.

So far it's pretty much been trying to find optimal ways of iterating
over things which is um, a thing but I didn't realize it was such
a big deal.

It feels like doing things in C would be so much faster but also
no one like working with C...

For now I'm mostly just trying to remember all the various structures
and ways of iterating over them.