Just calculate expected uniform values
I was talking to a friend about the problem and he mentioned
just "calculating all the expected values and checking if it's in
the range."

At first I thought it was a crazy idea b/c we don't acutally look
at any particular values in A or B but when thinking about it, for
any given power of 10, there's only 9 max uniform values.

So you could just "calulate" expected things if you figured out
the powers of 10 with a "max uniform value" within the range. Then
you only have to compare the inital pow of 10 and final pow of 10,
which is actually a pretty small amount.

This saves a TON on calculating (altho I think it's kinda cheating).

I broke the meta puzzle page tho somehow and it won't let me submit.

Seems to work for my test cases tho!

Thanks for the idea Sterling!

Co-authored-by: Sterling Hirsh <sterlinghirsh@users.noreply.github.com>